backend/requirements.txt
fastapi>=0.95
uvicorn>=0.22
pydantic>=1.10
azure-cosmos>=4.3.0
pytest>=7.0
pytest-mock>=3.0
pytest-asyncio>=0.22
python-jose>=3.3.0
passlib[bcrypt]>=1.7.4
